Added restrictions

+++ src/main/java/org/springframework/samples/petclinic/web/InvitationController.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.Invitation;
 import org.springframework.samples.petclinic.model.InvitationStatus;
 import org.springframework.samples.petclinic.model.Invitations;
+import org.springframework.samples.petclinic.model.JamStatus;
 import org.springframework.samples.petclinic.model.Team;
 import org.springframework.samples.petclinic.model.User;
 import org.springframework.samples.petclinic.service.InvitationService;
@@ -58,7 +59,9 @@ public class InvitationController {
 	@GetMapping(value = "/jams/{jamId}/teams/{teamId}/invitations/new")
 	public String initCreationForm(@PathVariable("teamId") final int teamId, final Map<String, Object> model)
 			throws Exception {
-		if (!this.teamService.findIsMemberOfTeamByTeamIdAndUsername(teamId, UserUtils.getCurrentUsername())) {
+		Team team = this.teamService.findTeamById(teamId);
+		if (!this.teamService.findIsMemberOfTeamByTeamIdAndUsername(teamId, UserUtils.getCurrentUsername())
+				|| team.getJam().getStatus() != JamStatus.INSCRIPTION) {
 			throw new Exception();
 		}
 
@@ -69,11 +72,12 @@ public class InvitationController {
 	@PostMapping(value = "/jams/{jamId}/teams/{teamId}/invitations/new")
 	public String processCreationForm(@Valid final Invitation invitation, final BindingResult result,
 			@PathVariable("teamId") final int teamId) throws Exception {
-		if (!this.teamService.findIsMemberOfTeamByTeamIdAndUsername(teamId, UserUtils.getCurrentUsername())) {
+		Team team = this.teamService.findTeamById(teamId);
+		if (!this.teamService.findIsMemberOfTeamByTeamIdAndUsername(teamId, UserUtils.getCurrentUsername())
+				|| team.getJam().getStatus() != JamStatus.INSCRIPTION) {
 			throw new Exception();
 		}
 
-		Team team = this.teamService.findTeamById(teamId);
 		String toUsername = invitation.getTo().getUsername();
 		User toUser = this.userService.findByUsername(toUsername);
 
@@ -84,7 +88,7 @@ public class InvitationController {
 				result.rejectValue("to.username", "pendingInvitation", "There's a pending invitation yet");
 			}
 			if (this.teamService.findIsMemberOfTeamByTeamIdAndUsername(teamId, toUsername)) {
-				result.rejectValue("to.username", "memberInvitation", "This user is a member of the team");
+				result.rejectValue("to.username", "isMember", "This user is a member of the team");
 			}
 		}
 
@@ -99,20 +103,31 @@ public class InvitationController {
 	}
 
 	@GetMapping(value = "/jams/{jamId}/teams/{teamId}/invitations/{invitationId}/delete")
-	public String initDeleteForm(@PathVariable("invitationId") final int invitationId, final ModelMap model) {
+	public String initDeleteForm(@PathVariable("invitationId") final int invitationId, final ModelMap model,
+			@PathVariable("teamId") final int teamId) throws Exception {
+		Team team = this.teamService.findTeamById(teamId);
 		Invitation invitation = this.invitationService.findInvitationById(invitationId);
-		model.remove("invitation", invitation);
+
+		if (invitation == null || invitation.getStatus() != InvitationStatus.PENDING
+				|| !this.teamService.findIsMemberOfTeamByTeamIdAndUsername(teamId, UserUtils.getCurrentUsername())
+				|| team.getJam().getStatus() != JamStatus.INSCRIPTION) {
+			throw new Exception();
+		}
+
 		this.invitationService.deleteInvitation(invitation);
+
 		return "redirect:/jams/{jamId}/teams/{teamId}";
 	}
 
 	@GetMapping(value = "/invitations/{invitationId}/accept")
 	public String processAccept(@PathVariable("invitationId") final int invitationId, final ModelMap model)
 			throws Exception {
-		// Comprobar que esta invitacion es para ese usuario
 		Invitation invitation = this.invitationService.findInvitationById(invitationId);
 		String currentUsername = UserUtils.getCurrentUsername();
-		if (!invitation.getTo().getUsername().equals(currentUsername)) {
+
+		if (invitation == null || invitation.getStatus() != InvitationStatus.PENDING
+				|| !invitation.getTo().getUsername().equals(currentUsername)
+				|| invitation.getFrom().getJam().getStatus() != JamStatus.INSCRIPTION) {
 			throw new Exception();
 		}
 
@@ -130,16 +145,18 @@ public class InvitationController {
 		team.getMembers().add(usuario);
 		this.teamService.saveTeam(team);
 
-		return "redirect:/invitations";
+		return "redirect:/jams/" + invitation.getFrom().getJam().getId() + "/teams/" + invitation.getFrom().getId();
 	}
 
 	@GetMapping(value = "/invitations/{invitationId}/reject")
 	public String processReject(@PathVariable("invitationId") final int invitationId, final ModelMap model)
 			throws Exception {
-		// Comprobar que esta invitacion es para ese usuario
 		Invitation invitation = this.invitationService.findInvitationById(invitationId);
 		String currentUsername = UserUtils.getCurrentUsername();
-		if (!invitation.getTo().getUsername().equals(currentUsername)) {
+
+		if (invitation == null || invitation.getStatus() != InvitationStatus.PENDING
+				|| !invitation.getTo().getUsername().equals(currentUsername)
+				|| invitation.getFrom().getJam().getStatus() != JamStatus.INSCRIPTION) {
 			throw new Exception();
 		}
